Update content URL to explicitely discard cache (at least for the page)
cf #7:

I think that was answered by Will Law in an email: "The account you are using is [a] test asset storage account, It is designed for very high volume delivery of video assets that don’t change very often (like test assets). [...] if you don’t supply the query arg, then it will give you the cached version. Bypassing the cache with the query arg does not then pull that object in to cache ;("