Update database.json
We would need to set that on the Akamai endpoint: https://dash-large-files.akamaized.net

For instance, for https://dash.akamaized.net/akamai/bbb_30fps/bbb_30fps.mpd it is set in the response headers:
access-control-allow-origin: *.

This does not seem to be the case for https://dash-large-files.akamaized.net. Just changing the URL to https://dash.akamaized.net/WAVE/vectors/cfhd_sets/12.5_25_50/t1/2023-09-01/stream.mpd seems to fix the issue.